Started spa pricing separation
started the structure for finding spa rebates for the customer.
Currently trying to think of the best way to handle finding what the
fulfillment is for that specific line item. Due to the duplicate line
items from customer purchases I don't want to have to keep reading in
data and finding the proper fulfillment. I think the best approach may
be to read all of the spa data at the start of the program and store it
efficiently so that it would be easy to retrieve when needed