Altered enVisionPartnerListClass
Altered the class to  account for a change in the structure of the
program, where there will be a global variable in the program itself to
read in the settings and get either default price levels or custom price
levels. The PriceLevels are being used to correlate enVision Status set
in GP. IE in the query enVision Status will return a number like '2'
that means the customer is of 'Select' enVision status, as told so by
GP. Will also check this against the spreadsheet containing envision
customers with their status'.

Keep in mind that the enVision Price Level as dictated in the
spreadsheet we get from Epson might be different that the GP definition
so it will be required that there be some form of mapping lets say
'Plus' to 'PlusPremier'