Finished SpaList class; moved itemFulfillmentClass
Finished SpaList class - the way you use it is:
SpaList sl = new SpaList();
bool = sl.initializeSpaList(filePath);

And then check to see if the spa list was initialized using the return
boolean value.

The structure of the Spa List workbook is as follows -
2 sheets in the workbook. The first sheet contains tables one column
wide of customers which will be named "spaCustomers_tableName". The
second sheet contains tables three columns wide of items that belong to
a specific spa. These tables will be named "tableName" and that is how
the items will be matched up with the corresponding customers.

Created itemFulfillmentClass's own file, to keep the workflow clean and
since it will be used in other classes besides spaList I thought it
would be best to separate it.