Added settings and finished price list class
Added settings for the column numbers in the price list. Will add later
an option to change the settings so if the column numbers every change
it can be an easy fix.

Also, reading in all rows from the price list and adding them to the
class list item. Need to ask Alan about the duplicate values in the
epson price list, and whether or not those can be removed and kept with
the latest. Otherwise will have to think of a good method of excluding
old values so there are no duplicates.