Started adding forms for UI
Removed custom user settings file, as I will not be using that, and
instead using Visual Studios UI.

Also, began adding the UI for the report form. Next step is to complete
the UI and PartnerList classes. Will continue testing multi-threading
and UI updates with a progress bar so the user knows what is going on