Multithreading and UI Progress
Completed the progress bar and the program layout for multithreading and
initializing the input spreadsheets while updating a progress bar to
inform the user that the program is running.

Currently working on - finishing the settings form which will be used to
set the file locations of the spreadsheets to correlate the report
against. These file locations will be saved into the application
settings of the program, to later be retrieved by the program and used.

Still need to add a 'save' and 'cancel' button to the form as well as
handle the file browsing and getting the filepaths.

Need to test the application settings in a separate program to confirm
that the syntax I am using is correct.